Move Chalk conversion code to its own module
